Update README.md
This Netflix Dashboard project showcases the real-world application of Excel for business intelligence and data storytelling. From budget and revenue analysis to regional content performance, this dashboard mimics the kind of reporting tools used by streaming services and media analysts. By combining data analysis, visualization, and interactivity, it provides users with a clear understanding of what content performs best and where.

Whether you're preparing for a data analyst role, curating a portfolio, or exploring Netflix trends, this dashboard stands out as a polished and insightful tool to analyze entertainment data dynamically. It's not just numbers—it's a story told with visuals, filters, and a Netflix-themed aesthetic that makes insights both easy and enjoyable to consume.